+++ deno.json
@@ -2,7 +2,7 @@
   "name": "@fcrozatier/monarch",
   "version": "2.6.0",
   "exports": {
-    ".": "./src/index.ts",
+    ".": "./src/main.ts",
     "./common": "./examples/common.ts",
     "./html": "./exports/html.ts",
     "./csv": "./exports/csv.ts"
