+++ examples/common.ts
@@ -10,7 +10,7 @@ import {
   result,
   sepBy,
   sequence,
-} from "../src/index.ts";
+} from "../src/main.ts";
 import { updatePosition } from "../src/utilities.ts";
 
 /**
