+++ lib/primitives/zero.ts
@@ -1,0 +1,12 @@
+import { createParser, type Parser } from "../../src/index.ts";
+
+/**
+ * The always failing parser
+ *
+ * It is the unit of alternation and plus, and also is an absorbing element of bind
+ */
+export const zero: Parser<never> = createParser((_, position) => ({
+  success: false,
+  message: "",
+  position,
+}));
